+++ .idea/misc.xml
@@ -3,7 +3,10 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/Asus/AndroidStudioProjects/Helpers_Around_You/app/src/main/res/layout/activity_login.xml" value="0.16666666666666666" />
         <entry key="..\:/Users/Asus/AndroidStudioProjects/Helpers_Around_You/app/src/main/res/layout/activity_main.xml" value="0.1" />
+        <entry key="..\:/Users/Asus/AndroidStudioProjects/Helpers_Around_You/app/src/main/res/layout/activity_splash_screen.xml" value="0.13451086956521738" />
+        <entry key="..\:/Users/Asus/AndroidStudioProjects/Helpers_Around_You/app/src/main/res/layout/activity_verify_otp.xml" value="0.13451086956521738" />
         <entry key="..\:/Users/Shruti/AndroidStudioProjects/HelpersAroundYou/app/src/main/res/drawable/edit_text_bg.xml" value="0.26615384615384613" />
         <entry key="..\:/Users/Shruti/AndroidStudioProjects/HelpersAroundYou/app/src/main/res/layout/activity_login.xml" value="0.264" />
         <entry key="..\:/Users/Shruti/AndroidStudioProjects/HelpersAroundYou/app/src/main/res/layout/activity_main.xml" value="0.25416666666666665" />
@@ -13,7 +16,7 @@
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="JDK" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="11" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
