--- app/src/main/java/com/trackme/helpersaroundyou/MainActivity.java
@@ -1,25 +1,0 @@
-package com.trackme.helpersaroundyou;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.os.Handler;
-
-public class MainActivity extends AppCompatActivity {
-    private final int SPLASH_TIME_OUT=4000;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        //this is splash screen
-
-        new Handler().postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                Intent intent=new Intent(MainActivity.this,Login.class);
-                startActivity(intent);
-            }
-        },SPLASH_TIME_OUT);
-    }
-}
