+++ app/src/main/java/com/trackme/helpersaroundyou/SplashScreen.java
@@ -2,13 +2,22 @@ package com.trackme.helpersaroundyou;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.os.Handler;
 
 public class SplashScreen extends AppCompatActivity {
-
+    private final int SPLASH_TIME_OUT=4000;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash_screen);
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent intent=new Intent(SplashScreen.this,Login.class);
+                startActivity(intent);
+            }
+        },SPLASH_TIME_OUT);
     }
 }
